Consolidate server-side-graphql-client code
