+++ .changeset/eight-jeans-raise.md
@@ -1,0 +1,5 @@
+---
+'@keystone-next/keystone': patch
+---
+
+Consolidate the code from the `@keystone-next/server-side-graphql-client-legacy` package into the main package.
