--- .changeset/hot-poets-buy.md
@@ -1,5 +1,0 @@
----
-'@keystone-next/server-side-graphql-client-legacy': patch
----
-
-Updated types to be more correct.
