+++ package.json
@@ -123,7 +123,6 @@
       "packages/test-utils",
       "packages/utils",
       "packages/adapter-prisma",
-      "packages/server-side-graphql-client",
       "packages-next/*",
       "design-system/packages/*"
     ],
