+++ packages-next/keystone/package.json
@@ -31,7 +31,6 @@
     "@keystone-next/adapter-prisma-legacy": "^6.0.0",
     "@keystone-next/admin-ui": "^14.0.0",
     "@keystone-next/fields": "^7.0.0",
-    "@keystone-next/server-side-graphql-client-legacy": "^4.0.0",
     "@keystone-next/types": "^17.0.0",
     "@keystone-next/utils-legacy": "^9.0.0",
     "@preconstruct/next": "^3.0.0",
