--- packages/server-side-graphql-client/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@keystone-next/server-side-graphql-client-legacy",
-  "description": "A library for running server-side graphQL queries and mutations in Keystone.",
-  "version": "4.0.0",
-  "author": "The KeystoneJS Development Team",
-  "license": "MIT",
-  "homepage": "https://github.com/keystonejs/keystone",
-  "repository": "https://github.com/keystonejs/keystone/tree/master/packages/server-side-graphql-client",
-  "main": "dist/server-side-graphql-client-legacy.cjs.js",
-  "module": "dist/server-side-graphql-client-legacy.esm.js",
-  "engines": {
-    "node": ">=v12.13.1"
-  },
-  "dependencies": {
-    "@keystone-next/types": "^17.0.0",
-    "graphql": "^15.5.0"
-  },
-  "devDependencies": {
-    "@keystone-next/test-utils-legacy": "*"
-  }
-}
