--- packages/server-side-graphql-client/src/index.ts
@@ -1,10 +1,0 @@
-export {
-  getItem,
-  getItems,
-  createItems,
-  createItem,
-  updateItem,
-  updateItems,
-  deleteItem,
-  deleteItems,
-} from './server-side-graphql-client';
