+++ tests/api-tests/package.json
@@ -15,7 +15,6 @@
     "@keystone-next/auth": "*",
     "@keystone-next/fields": "*",
     "@keystone-next/keystone": "*",
-    "@keystone-next/server-side-graphql-client-legacy": "*",
     "cookie-signature": "^1.1.0",
     "cuid": "^2.1.8",
     "globby": "^11.0.3",
